Load policy checkpoints directly in bf16 to halve peak load memory
Gr00tPolicy previously called AutoModel.from_pretrained with no torch_dtype,
so the bf16 checkpoint was materialized in fp32 (~2x the steady-state size)
before being cast back down by the .to(device, dtype) call. Passing
torch_dtype=torch.bfloat16 skips the transient fp32 copy; the final state is
bit-identical because bf16 -> fp32 -> bf16 is a lossless round-trip, and the
existing .to() is kept to normalize any stragglers exactly as before.

Measured on a 100M-param synthetic bf16 checkpoint (fresh process per mode,
ru_maxrss): peak RSS 1281 MiB -> 708 MiB; the ~578 MiB delta matches the
removed fp32 materialization. For the ~3B production checkpoint this avoids
a ~10 GB transient, which is most of the gap between the documented 16 GB
inference floor and the model's ~6 GB steady-state footprint.

Tests pin both the wiring (from_pretrained receives torch_dtype, .to() and
eval() still applied) and the guarantee (state dict and produced actions are
bitwise identical to the legacy fp32-load-then-cast path).

+++ gr00t/policy/gr00t_policy.py
@@ -105,8 +105,13 @@ class Gr00tPolicy(BasePolicy):
             embodiment_tag = EmbodimentTag.resolve(embodiment_tag)
         model_dir = Path(model_path)
 
-        # Load the pretrained model and move to target device with bfloat16 precision
-        model = AutoModel.from_pretrained(model_dir)
+        # Load the pretrained model and move to target device with bfloat16 precision.
+        # torch_dtype=torch.bfloat16 materializes the checkpoint (stored in bf16)
+        # directly in bf16 instead of upcasting to fp32 first, halving peak host
+        # memory and load time. The final state is unchanged: bf16 -> fp32 -> bf16
+        # is a lossless round-trip, and the .to() below still normalizes any
+        # stragglers (e.g. non-persistent buffers) exactly as before.
+        model = AutoModel.from_pretrained(model_dir, torch_dtype=torch.bfloat16)
         model.eval()  # Set model to evaluation mode
         model.to(device=device, dtype=torch.bfloat16)
         self.model = model
